pyproject.toml
[tool.poetry]
name = "pluggy-test-python"
version = "0.1.0"
description = ""
authors = ["James Bristow <jbris@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pluggy = "^1.4.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
